refactor: rename components and create index file for them
