+++ src/components/Footer.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import './Footer.css';
+
+import './footer.css';
 
 export default function Footer() {
     return (
