+++ src/components/Header.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import './Header.css';
 import { Link } from 'react-router-dom';
+
+import './header.css';
 
 export default function Header() {
     return (
