+++ src/components/index.js
@@ -1,0 +1,7 @@
+export { default as Footer } from './footer';
+export { default as Header } from './header';
+export { default as Pagination } from './pagination';
+export { default as PostCard } from './post-card';
+export { default as PostsList } from './posts-list';
+export { default as PagePost } from './page-post';
+export { default as PageIndex } from './page-index';
